Expence-Manager-Backend/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2025.7.9
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.6.1
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
fonttools==4.58.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
motor==3.3.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
pyasn1==0.6.1
pydantic>=2.7.0,<3
pydantic_core>=2.14.0,<3
PyJWT==2.8.0
pymongo==4.6.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.23.2
